Refactor: #20 - ItemQuantityLogRepository Querydsl로 변경

+++ src/main/java/com/house/item/domain/QuantityLogSumByDate.java
@@ -9,5 +9,5 @@ import lombok.ToString;
 @ToString
 public class QuantityLogSumByDate {
 	private int date;
-	private long sum;
+	private int sum;
 }
